Move code from traversal to dag_algo
